Many changes to make the communication work between the Arduino Uno and the ProMini. the data structure is now optimised so that interrupt routine can be faster - it only needs a delay of 1ms between successive sends from the Uno. Also changed the ProMini so that it can require the Uno to start afterwards so that it does not get out of sync with control and data bytes. Before its first control byte it requires a perfect sequence of six bytes to arrive. This is achieved at any time by restarting the Uno. After this the systems are lock stepped on control and data bytes.
Also added a simple CommsTest module for the Uno that can be used to produce nice sequences of commands for the ProMini without the need to run the car in a hostile environment.

The Arduino structure in the ProMini code is the key to understanding how this all works.  It is straightforward but does use some indirect arrays for efficiency.
We copy data out of the receive data buffer to a user buffer to prevent the case that the user asks for data only when a data section is half way through (code arrived but not yet all data).  Data can still be missed but this is unimportant as the Uno is behaving like a sensor array for the ProMini.

Now the Uno can only send four kinds of messages.
One expected advantage of the new two machine approach is that we should be able to stop the car during three point turn manoeuvres.