BUG FIXES and Readability improvements -------------------------------------- RFID:process_rfid A line had accidentally been deleted.  Whne a new RFID tag is found it not only changes the current status of the car, but also needs to save the RFID string for use in other methods. Also allowed different speed ranges (30-70) works now for speed changes.
void ProMini::setStopped
Changed the zero speed byte to uni formally use the toSpeedByte method for readability.

ProMini::setTurn
Set remembered last speed to zero because the Turn code in the ProMini finishes at zero speed.
Changed the code so that is actually uses the parameters from the Android app.
Now waits (up to 1sec) for a HIGH isMoving pin to indicate that a move has started, rather than just waiting a default 1sec.

ProMini::send
Changed the delay time to 300usec rather that a few milliseconds as this is enough for the ISR on the ProMini to complete (tested).

lineFollow.h
Changed MIN_ANGLE, NEUTRAL_ANGLE and MAX_ANGLE to be ints (rather than unsigned) to correct calculation errors.

setup:
Now initialised three point turn constants to reasonable values so that something will happen even if they have not been set from Android.
Also set initial desired speed much lower.

-----------
Also changed all printed strings everywhere to begin with the code # and to end with the standard END_STRING.  To facilitate this moved the END_STRING to be a manifest constant in the header file.