LineTrack and AutoTrack are now the same except that LineTrack is standalone. So, LineTrack just calibrates until done and then tracks a line AutoTrack calibrates until done whenever the Follow Pin says it should! otherwise it follows.
the calibration is to find two near vertical lines about the right distance apart and to use the gap between to alter the current find_blobs threshold.
When find_blobs reliably finds just one small blob we are done.

Of course this means that the file Calibrate.py is redundant.